+++ AppboyKit/ABKFeedController+InnerMethods.h
@@ -1,0 +1,8 @@
+#import "ABKFeedController.h"
+#import "ABKPersistentData.h"
+
+@interface ABKFeedController (InnerMethods)
+
+- (instancetype)initWithPersistentData:(ABKPersistentData *)persistentData;
+
+@end