+++ AppboyKit/headers/AppboyKitLibrary/ABKNavigationBar.h
@@ -1,8 +1,0 @@
-#import <UIKit/UIKit.h>
-
-/*
- * Braze Public API: ABKNavigationBar
- */
-@interface ABKNavigationBar : UINavigationBar
-
-@end