+++ AppboyKit/headers/AppboyKitLibrary/ABKThemableFeedNavigationBar.h
@@ -1,8 +1,0 @@
-#import "ABKNavigationBar.h"
-
-/*
- * Appboy Public API: ABKThemableFeedNavigationBar
- */
-@interface ABKThemableFeedNavigationBar : ABKNavigationBar
-
-@end