+++ AppboyUI/ABKFeedViewController/FeedViewController/ABKNewsFeedTableViewController.m
@@ -1,0 +1,253 @@
+#import "ABKNewsFeedTableViewController.h"
+#import "ABKNFBannerCardCell.h"
+#import "UIImageView+WebCache.h"
+#import "ABKNFCaptionedMessageCardCell.h"
+#import "ABKNFClassicCardCell.h"
+#import "ABKNFCrossPromotionCardCell.h"
+#import "ABKUIUtils.h"
+#import "ABKFeedWebViewController.h"
+#import "ABKUIURLUtils.h"
+#import <SDWebImage/SDWebImagePrefetcher.h>
+
+@implementation ABKNewsFeedTableViewController
+
+#pragma mark - Initialization
+
+- (instancetype)init {
+  UIStoryboard *st = [UIStoryboard storyboardWithName:@"ABKNewsFeedCardStoryboard"
+                                               bundle:[NSBundle bundleForClass:[ABKNewsFeedTableViewController class]]];
+  ABKNewsFeedTableViewController *nf = [st instantiateViewControllerWithIdentifier:@"ABKNewsFeedTableViewController"];
+  self = nf;
+  return self;
+}
+
+- (instancetype)initWithCoder:(NSCoder *)aDecoder {
+  self = [super initWithCoder:aDecoder];
+  if (self) {
+    [self setUp];
+  }
+  return self;
+}
+
+- (void)setUp {
+  _categories = ABKCardCategoryAll;
+  _cacheTimeout = 60.0;
+  
+  [[NSNotificationCenter defaultCenter] addObserver:self
+                                           selector:@selector(feedUpdated:)
+                                               name:ABKFeedUpdatedNotification
+                                             object:nil];
+}
+
+# pragma mark - View Controller Life Cycle Methods
+
+- (void)viewDidLoad {
+  [super viewDidLoad];
+  
+  self.cards = [[Appboy sharedInstance].feedController getCardsInCategories:self.categories];
+  
+  self.tableView.rowHeight = UITableViewAutomaticDimension;
+  self.tableView.estimatedRowHeight = 160;
+  
+  [self requestNewCardsIfTimeout];
+  [self cacheAllCardImages];
+  
+  self.emptyFeedLabel.text = [self localizedAppboyFeedString:@"Appboy.feed.no-card.text"];
+}
+
+- (void)viewWillAppear:(BOOL)animated {
+  [super viewWillAppear:animated];
+  [self updateAndDisplayCardsFromCache];
+  self.constraintWarningValue =
+    [[NSUserDefaults standardUserDefaults] valueForKey:@"_UIConstraintBasedLayoutLogUnsatisfiable"];
+  [[NSUserDefaults standardUserDefaults] setValue:@(NO) forKey:@"_UIConstraintBasedLayoutLogUnsatisfiable"];
+}
+
+- (void)viewWillDisappear:(BOOL)animated {
+  [super viewWillDisappear:animated];
+  [[NSUserDefaults standardUserDefaults] setValue:self.constraintWarningValue forKey:@"_UIConstraintBasedLayoutLogUnsatisfiable"];
+}
+
+- (void)viewWillTransitionToSize:(CGSize)size
+       withTransitionCoordinator:(id<UIViewControllerTransitionCoordinator>)coordinator {
+  [super viewWillTransitionToSize:size withTransitionCoordinator:coordinator];
+  [self.tableView reloadData];
+}
+
+- (void)dealloc {
+  [[NSNotificationCenter defaultCenter] removeObserver:self];
+}
+
+#pragma mark - Update And Display Cached Cards
+
+- (IBAction)refreshNewsFeed:(UIRefreshControl *)sender {
+  [[Appboy sharedInstance] requestFeedRefresh];
+}
+
+- (void)requestNewCardsIfTimeout {
+  NSTimeInterval passedTime = [[Appboy sharedInstance].feedController.lastUpdate timeIntervalSinceNow];
+  if (passedTime > self.cacheTimeout) {
+    [[Appboy sharedInstance] requestFeedRefresh];
+  }
+}
+
+- (void)feedUpdated:(NSNotification *)notification {
+  BOOL isSuccessful = [notification.userInfo[ABKFeedUpdatedIsSuccessfulKey] boolValue];
+  if (isSuccessful) {
+    [self updateAndDisplayCardsFromCache];
+  }
+  [self.refreshControl endRefreshing];
+}
+
+- (void)updateAndDisplayCardsFromCache {
+  self.cards = [[Appboy sharedInstance].feedController getCardsInCategories:self.categories];
+  if (self.cards == nil || self.cards.count == 0) {
+    [self hideTableViewAndShowViewInHeader:self.emptyFeedView];
+  } else {
+    [self showTableViewAndHideHeaderViews];
+  }
+  [self.tableView reloadData];
+}
+
+- (void)hideTableViewAndShowViewInHeader:(UIView *)view {
+  view.hidden = NO;
+  view.frame = self.view.bounds;
+  [view layoutIfNeeded];
+  self.tableView.sectionHeaderHeight = self.tableView.frame.size.height;
+  self.tableView.tableHeaderView = view;
+  self.tableView.scrollEnabled = NO;
+}
+
+- (void)showTableViewAndHideHeaderViews {
+  self.emptyFeedView.hidden = YES;
+  self.tableView.tableHeaderView = nil;
+  self.tableView.sectionHeaderHeight = 0;
+  self.tableView.scrollEnabled = YES;
+}
+
+#pragma mark - Configuration Update
+
+- (void)setDisableUnreadIndicator:(BOOL)disableUnreadIndicator {
+  if (disableUnreadIndicator != _disableUnreadIndicator) {
+    _disableUnreadIndicator = disableUnreadIndicator;
+    [self updateAndDisplayCardsFromCache];
+  }
+}
+
+- (void)setCategories:(ABKCardCategory)categories {
+  if (categories != _categories) {
+    _categories = categories;
+    [self updateAndDisplayCardsFromCache];
+  }
+}
+
+#pragma mark - Table view data source
+
+- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
+    return 1;
+}
+
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+    return self.cards.count;
+}
+
+- (void)tableView:(UITableView *)tableView
+  willDisplayCell:(UITableViewCell *)cell
+forRowAtIndexPath:(NSIndexPath *)indexPath {
+  ABKCard *card = self.cards[indexPath.row];
+  [card logCardImpression];
+}
+
+- (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
+  ABKCard *card = self.cards[indexPath.row];
+  ABKNFBaseCardCell *cell = [ABKNFBaseCardCell dequeueCellFromTableView:tableView
+                                                           forIndexPath:indexPath
+                                                                forCard:card];
+  [cell applyCard:card];
+  cell.hideUnreadIndicator = self.disableUnreadIndicator;
+  if ([cell isKindOfClass:[ABKNFCrossPromotionCardCell class]]) {
+    ((ABKNFCrossPromotionCardCell *)cell).actionDelegate = self;
+  }
+  return cell;
+}
+
+- (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
+  ABKCard *card = self.cards[indexPath.row];
+  [self handleCardClick:card];
+}
+
+#pragma mark - Card Click Actions
+
+- (void)handleCardClick:(ABKCard *)card {
+  [card logCardClicked];
+  
+  NSURL *cardURL = [ABKUIURLUtils getEncodedURIFromString:card.urlString];
+   if ([ABKUIURLUtils URL:cardURL shouldOpenInWebView:card.openUrlInWebView]) {
+     [self openURLInWebView:cardURL];
+   } else {
+     [ABKUIURLUtils openURLWithSystem:cardURL];
+   }
+}
+
+- (void)openURLInWebView:(NSURL *)url {
+  ABKFeedWebViewController *webViewController = [[ABKFeedWebViewController alloc] init];
+  webViewController.url = url;
+  webViewController.showDoneButton = self.navigationItem.rightBarButtonItem != nil;
+  [self.navigationController pushViewController:webViewController animated:YES];
+}
+
+- (void)openItunesStoreProductWithId:(NSNumber *)productId url:(NSURL *)url {
+  Class SKStoreProductViewControllerClass = NSClassFromString(@"SKStoreProductViewController");
+  if (SKStoreProductViewControllerClass != nil) {
+    SKStoreProductViewController *productViewController = [[SKStoreProductViewControllerClass alloc] init];
+    productViewController.delegate = self;
+
+    [self presentViewController:productViewController animated:YES completion:nil];
+    [productViewController loadProductWithParameters:@{SKStoreProductParameterITunesItemIdentifier : productId}
+                                     completionBlock:^(BOOL result, NSError *error) {
+                                       if (error) {
+                                         NSLog(@"%@%@:%@", self, NSStringFromSelector(_cmd),[error localizedDescription]);
+                                       }
+                                     }
+     ];
+  } else {
+    [self openURLInWebView:url];
+  }
+}
+
+- (void)productViewControllerDidFinish:(SKStoreProductViewController *)viewController {
+  [self dismissViewControllerAnimated:YES completion:nil];
+}
+
+#pragma mark - Image Caching
+
+- (void)cacheAllCardImages {
+  NSMutableArray *images = [NSMutableArray arrayWithCapacity:self.cards.count];
+  for (ABKCard *card in self.cards) {
+    if ([card respondsToSelector:@selector(image)]) {
+      NSString *imageUrlString = [[card performSelector:@selector(image)] copy];
+      NSURL *imageUrl = [ABKUIURLUtils getEncodedURIFromString:imageUrlString];
+      if ([ABKUIUtils objectIsValidAndNotEmpty:imageUrl]) {
+        [images addObject:imageUrl];
+      }
+    }
+  }
+  [[SDWebImagePrefetcher sharedImagePrefetcher] prefetchURLs:images];
+}
+
+# pragma mark - Utility Methods
+
++ (instancetype)getNavigationFeedViewController {
+  UIStoryboard *st = [UIStoryboard storyboardWithName:@"ABKNewsFeedCardStoryboard"
+                                               bundle:[NSBundle bundleForClass:[ABKNewsFeedTableViewController class]]];
+  ABKNewsFeedTableViewController *nf = [st instantiateViewControllerWithIdentifier:@"ABKNewsFeedTableViewController"];
+  return nf;
+}
+
+- (NSString *)localizedAppboyFeedString:(NSString *)key {
+  return [ABKUIUtils getLocalizedString:key
+                         inAppboyBundle:[NSBundle bundleForClass:[ABKNewsFeedTableViewController class]]
+                                  table:@"AppboyFeedLocalizable"];
+}
+
+@end