+++ AppboyUI/ABKFeedViewController/FeedViewController/AppboyNewsFeed.h
@@ -1,0 +1,12 @@
+// Braze News Feed View Controllers
+#import "ABKFeedWebViewController.h"
+#import "ABKNewsFeedTableViewController.h"
+#import "ABKNewsFeedViewController.h"
+
+// Braze News Feed Cells
+#import "ABKNFBannerCardCell.h"
+#import "ABKNFBaseCardCell.h"
+#import "ABKNFCaptionedMessageCardCell.h"
+#import "ABKNFClassicCardCell.h"
+#import "ABKNFCrossPromotionCardActionDelegate.h"
+#import "ABKNFCrossPromotionCardCell.h"