+++ AppboyUI/ABKFeedViewController/FeedViewController/Cells/ABKNFCrossPromotionCardActionDelegate.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+
+@protocol ABKNFCrossPromotionCardActionDelegate <NSObject>
+
+- (void)openItunesStoreProductWithId:(NSNumber *)productId url:(NSURL *)url;
+
+@end