+++ AppboyUI/ABKUIUtils/ABKUIUtils.h
@@ -8,5 +8,8 @@
 + (Class)getSDWebImageProxyClass;
 + (Class)getModalFeedViewControllerClass;
 + (BOOL)isiPhoneX;
++ (UIImage *)getImageWithName:(NSString *)name
+                         type:(NSString *)type
+               inAppboyBundle:(NSBundle *)appboyBundle;
 
 @end