+++ Samples/NewsFeed/BrazeNewsFeedSample/BrazeNewsFeedSample/CustomFeedTableViewController.h
@@ -1,0 +1,5 @@
+#import "AppboyNewsFeed.h"
+
+@interface CustomFeedTableViewController : ABKNewsFeedTableViewController
+
+@end