+++ Samples/NewsFeed/BrazeNewsFeedSample/BrazeNewsFeedSample/ViewController.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+
+@interface ViewController : UIViewController
+
+@property (weak, nonatomic) IBOutlet UISwitch *unreadIndicatorSwitch;
+
+@end